Add collector-gdb.sh script
This script essentially follows the steps described here:
https://github.com/stackrox/collector/blob/master/docs/how-to-start.md#debugging-on-a-k8s-deployment

But it does all the steps automatically, making it easier to jump into a
gdb debug session for collector.

The script also includes a hotreload mode, allowing for the collector
pod to run with a locally built binary, making debugging local changes
a lot faster. Bare in mind that this mode will only work on local
clusters like kind, k3s, etc...